Revert "Handles multiple input urls, output urls, background colours, model viewer args"
This reverts commit ed71df46f2b18b726a697b54326c32abc531705d.
