+++ .gitignore
@@ -94,7 +94,5 @@ package-lock.json
 
 output.jpg
 output.png
-output1.png
-output2.png
 test/screenshots
 .DS_Store
