+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@shopify/screenshot-glb",
-  "version": "2.0.0",
+  "version": "3.0.0",
   "publishConfig": {
     "access": "public"
   },
@@ -18,8 +18,6 @@
     "dev:version": "yarn build && node dist/cli.js -i ./test/fixtures/WaterBottle.glb -o output.jpg -@ 1.9 -w 1024 -h 2048 -q 1.00 -v -t 30000 -m \"exposure=0.92&camera-orbit=88.59deg 59.01deg 0.24m&camera-target=0.06m 0.04m 0m&environment-image=neutral\"",
     "dev:no-version": "yarn build && node dist/cli.js -i ./test/fixtures/WaterBottle.glb -o output.jpg -@ 100000.0 -w 1024 -h 2048 -q 1.00 -v -t 30000 -m \"exposure=0.92&camera-orbit=88.59deg 59.01deg 0.24m&camera-target=0.06m 0.04m 0m&environment-image=neutral\"",
     "dev:png": "yarn build && node dist/cli.js -i ./test/fixtures/WaterBottle.glb -o output.png -c \"rgba(0.0, 0.0, 0.0, 0.0)\" -w 1024 -h 2048 -q 1.00 -v -t 30000 -m \"exposure=0.92&camera-orbit=88.59deg 59.01deg 0.24m&camera-target=0.06m 0.04m 0m&environment-image=neutral\"",
-    "dev:multiple": "yarn build && node dist/cli.js -w 1024 -h 2048 -q 1.00 -v -t 30000 -i \"./test/fixtures/WaterBottle.glb,./test/fixtures/WaterBottle.glb\" -o \"output1.png,output2.png\" -c \"#FF00FF,#CAFE00\" -m \"exposure=0.92&camera-orbit=0deg+0deg+0.8m&camera-target=0.06m+0.04m+0m&environment-image=neutral,exposure=0.92&camera-orbit=88.59deg+59.01deg+0.24m&camera-target=0.02m+0.01m+0m&environment-image=neutral\"",
-    "dev:multiple-less-args": "yarn build && node dist/cli.js -w 1024 -h 2048 -q 1.00 -v -t 30000 -i \"./test/fixtures/WaterBottle.glb,./test/fixtures/WaterBottle.glb\" -o \"output1.png,output2.png\" -c \"#CAFE00\"",
     "dev:local-js": "yarn build && node dist/cli.js -i ./test/fixtures/WaterBottle.glb -o output.jpg --model_viewer_path ./test/fixtures/model-viewer.js -w 1024 -h 2048 -q 1.00 -v -t 30000 -m \"exposure=0.92&camera-orbit=88.59deg 59.01deg 0.24m&camera-target=0.06m 0.04m 0m&environment-image=neutral\""
   },
   "author": "Stephan Leroux <stephan.leroux@shopify.com>",
