+++ src/capture-screenshot.test.ts
@@ -48,27 +48,28 @@ describe('captureScreenshot', () => {
   const formatExtension = 'jpeg';
   const defaultParams = {
     modelViewerUrl,
-    inputUrls: [inputPath],
-    outputPaths: [outputPath],
-    backgroundColors: [backgroundColor],
-    modelViewerArgs: [],
+    inputPath,
+    outputPath,
     debug,
     quality,
     timeout,
     width,
     height,
+    backgroundColor,
     devicePixelRatio,
     formatExtension,
   };
   const htmlContent = '<div>some html</div>';
   let originalConsoleLog: typeof console.log;
   let mockPage;
+  let mockBrowser;
 
   beforeEach(() => {
     originalConsoleLog = console.log;
     console.log = jest.fn();
 
     mockPage = jest.requireMock('puppeteer').mock.page as Page;
+    mockBrowser = jest.requireMock('puppeteer').mock.browser as Browser;
 
     (htmlTemplate as jest.Mock).mockReturnValue(htmlContent);
     (performance.now as jest.Mock).mockReturnValue(0);
@@ -142,17 +143,16 @@ describe('captureScreenshot', () => {
       '🗺  Loading template to DOMContentLoaded (0.00s)',
       '🖌  Rendering screenshot of model (0.00s)',
       '🖼  Captured screenshot (0.00s)',
-      '🪂  Closed browser (0.00s)',
     ];
 
     await captureScreenshot({
       ...defaultParams,
     });
 
+    expect(console.log).toHaveBeenCalledTimes(expectedLogs.length);
     expectedLogs.forEach((log, i) => {
       expect(console.log).toHaveBeenNthCalledWith(i + 1, log);
     });
-    expect(console.log).toHaveBeenCalledTimes(expectedLogs.length);
   });
 
   test('handles evaluate error', async () => {
@@ -161,9 +161,7 @@ describe('captureScreenshot', () => {
       '🚀  Launched browser (0.00s)',
       '🗺  Loading template to DOMContentLoaded (0.00s)',
       '🖌  Rendering screenshot of model (0.00s)',
-      '❌  Evaluate error: Error: some error',
-      '❌  Closing browser because of error',
-      '🪂  Closed browser (0.00s)',
+      `❌  Evaluate error: ${error}`,
     ];
 
     mockPage.evaluate.mockResolvedValue(error);
@@ -172,11 +170,10 @@ describe('captureScreenshot', () => {
       ...defaultParams,
     });
 
+    expect(console.log).toHaveBeenCalledTimes(expectedLogs.length);
     expectedLogs.forEach((log, i) => {
       expect(console.log).toHaveBeenNthCalledWith(i + 1, log);
     });
-    expect(console.log).toHaveBeenCalledTimes(expectedLogs.length);
-
     expect(mockPage.screenshot).not.toHaveBeenCalled();
 
     mockPage.evaluate.mockResolvedValue(undefined);
