+++ src/html-template.ts
@@ -1,4 +1,6 @@
-export type AttributesObject = {[key: string]: any};
+import {getModelViewerUrl} from './get-model-viewer-url';
+
+type AttributesObject = {[key: string]: any};
 
 export interface TemplateRenderOptions {
   modelViewerUrl: string;
