--- src/prepare-app-options/model-viewer-attributes.ts
@@ -1,26 +1,0 @@
-import {AttributesObject} from '../html-template';
-import {REGEX_PARAM_SEPARATOR} from './param-separator';
-
-export function modelViewerAttributes(
-  model_viewer_attributes?: string,
-): AttributesObject[] {
-  if (!model_viewer_attributes) {
-    return [];
-  }
-
-  const modelViewerAttributes = model_viewer_attributes.split(
-    REGEX_PARAM_SEPARATOR,
-  );
-
-  return modelViewerAttributes.map((attributes) => {
-    const modelViewerArgs = {};
-
-    const params = new URLSearchParams(attributes);
-
-    params.forEach((value, key) => {
-      modelViewerArgs[key] = value;
-    });
-
-    return modelViewerArgs;
-  });
-}
