--- src/prepare-app-options/param-separator.ts
@@ -1,1 +1,0 @@
-export const REGEX_PARAM_SEPARATOR = /\s*,\s*/;
